+++ include/JoSIM/RelevantTrace.hpp
@@ -12,7 +12,7 @@ namespace JoSIM {
 
   class Matrix;
   class Input;
-  enum class StorageType : int {
+  enum class StorageType : int64_t {
     Voltage = 0, Phase = 1, Current = 2
   };
 
@@ -59,16 +59,16 @@ namespace JoSIM {
     int_o index2;
     int_o sourceIndex;
     int_o variableIndex;
-    int fIndex = -1;
+    int64_t fIndex = -1;
 
     RelevantTrace() :
       device(false) {};
   }; // class RelevantTrace
 
   void find_relevant_traces(Input& iObj, Matrix& mObj);
-  void handle_current(const std::string& s, Matrix& mObj, int fIndex);
+  void handle_current(const std::string& s, Matrix& mObj, int64_t fIndex);
   void handle_voltage_or_phase(
-    const std::string& s, bool voltage, Matrix& mObj, int fIndex);
+    const std::string& s, bool voltage, Matrix& mObj, int64_t fIndex);
 
 } // namespace JoSIM
