+++ include/JoSIM/TransmissionLine.hpp
@@ -51,22 +51,22 @@ namespace JoSIM {
 
   class TransmissionLine : public BasicComponent {
     private:
-    int hDepPos_ = 0;
+    int64_t hDepPos_ = 0;
     JoSIM::AnalysisType at_;
     public:
     NodeConfig nodeConfig2_;
     int_o posIndex2_, negIndex2_;
-    int currentIndex2_ = 0;
+    int64_t currentIndex2_ = 0;
     double n1_1_ = 0.0, n1_2_ = 0.0, n2_1_ = 0.0, n2_2_ = 0.0;
     double nk_1_ = 0.0, nk_2_ = 0.0, nk1_1_ = 0.0, nk1_2_ = 0.0, nk2_1_ = 0.0,
       nk2_2_ = 0.0;
-    int timestepDelay_ = 0;
+    int64_t timestepDelay_ = 0;
 
     TransmissionLine(
       const std::pair<tokens_t, string_o>& s, const NodeConfig& ncon,
       const std::optional<NodeConfig>& ncon2, const nodemap& nm,
       std::unordered_set<std::string>& lm, nodeconnections& nc,
-      const param_map& pm, const AnalysisType& at, const double& h, int& bi);
+      const param_map& pm, const AnalysisType& at, const double& h, int64_t& bi);
 
     void set_secondary_node_indices(
       const tokens_t& t, const nodemap& nm, nodeconnections& nc);
