+++ src/CCCS.cpp
@@ -26,7 +26,7 @@ CCCS::CCCS(
   const std::pair<tokens_t, string_o>& s, const NodeConfig& ncon,
   const std::optional<NodeConfig>& ncon2, const nodemap& nm,
   std::unordered_set<std::string>& lm, nodeconnections& nc,
-  const param_map& pm, int& bi) {
+  const param_map& pm, int64_t& bi) {
   // Check if the label has already been defined
   if (lm.count(s.first.at(0)) != 0) {
     Errors::invalid_component_errors(
