+++ src/CurrentSource.cpp
@@ -16,7 +16,7 @@ using namespace JoSIM;
 
 CurrentSource::CurrentSource(
   const std::pair<tokens_t, string_o>& s, const NodeConfig& ncon,
-  const nodemap& nm, std::unordered_set<std::string>& lm, const int& si) {
+  const nodemap& nm, std::unordered_set<std::string>& lm, const int64_t& si) {
   // Check if the label has already been defined
   if (lm.count(s.first.at(0)) != 0) {
     Errors::invalid_component_errors(
