+++ src/Resistor.cpp
@@ -30,9 +30,9 @@ using namespace JoSIM;
 Resistor::Resistor(
   const std::pair<tokens_t, string_o>& s, const NodeConfig& ncon,
   const nodemap& nm, std::unordered_set<std::string>& lm, nodeconnections& nc,
-  Input& iObj, Spread& spread, int& bi) {
+  Input& iObj, Spread& spread, int64_t& bi) {
   double spr = 1.0;
-  for (int i = 4; i < s.first.size(); ++i) {
+  for (int64_t i = 4; i < s.first.size(); ++i) {
     auto& t = s.first.at(i);
     if (t.rfind("SPREAD=", 0) == 0) {
       spr = parse_param(t.substr(7), iObj.parameters, s.second);
