+++ src/VCCS.cpp
@@ -35,7 +35,7 @@ VCCS::VCCS(
   const std::pair<tokens_t, string_o>& s, const NodeConfig& ncon,
   const std::optional<NodeConfig>& ncon2, const nodemap& nm,
   std::unordered_set<std::string>& lm, nodeconnections& nc,
-  const param_map& pm, int& bi, const AnalysisType& at, const double& h) {
+  const param_map& pm, int64_t& bi, const AnalysisType& at, const double& h) {
   at_ = at;
   // Set the label
   netlistInfo.label_ = s.first.at(0);
